+++ AntAlgorithms/ParallelOptimisationWithInheritance/AntSystemParallelOptimisationWithInheritance.cs
@@ -53,7 +53,7 @@ namespace ParallelOptimisationWithInheritance
             return nextColonyId;
         }
 
-        public double[] CalculateProbability(short interSectionId, int nextColony)
+        public decimal[] CalculateProbability(short interSectionId, int nextColony)
         {
             var antSystemFragment = AntSystemFragments[interSectionId];
