+++ AntAlgorithms/ParallelOptimisationWithInheritance/AspgParallelOptimisationWithInheritance.cs
@@ -31,7 +31,7 @@ namespace ParallelOptimisationWithInheritance
                         vertexIndex++)
                     {
                         var nextColony = antSystem.GetNextColony(interSectionIndex);
-                        double[] probability = antSystem.CalculateProbability(interSectionIndex, nextColony);
+                        decimal[] probability = antSystem.CalculateProbability(interSectionIndex, nextColony);
                         var chosenVertexIndex = Roulette(probability);
                         var chosenVertex = Graph.VerticesWeights[chosenVertexIndex];
                         antSystem.AddFreeVertexToTreil(interSectionIndex, nextColony, chosenVertex);
